polymarket-copy-bot/requirements.txt
web3==6.15.1
py-clob-client==0.18.0
python-dotenv==1.0.1
aiohttp==3.9.5
requests==2.31.0
